requirements.txt
openai
pandas
numpy
scikit-learn
matplotlib
seaborn
plotly
scipy
python-dotenv